Keep status of non-staged cherry-picked changes on build approve
A fix on build approve method.

When running 'build approve' after merging a branch to another
branch, an incorrect status change was fulfilled:
The status of cherry picked change (on targed branch) was changed
from new/abandoned to merged.

This occurred because cherry picked change had the same change-id as
one of the merged changes on source branch.

Fixes: QTQAINFRA-3411
Fixes: QTQAINFRA-3374
Change-Id: Ie04c5268b0196f5ff9e43618fe93c300865b9ea6
Reviewed-by: Jukka Jokiniva <jukka.jokiniva@qt.io>