Remove API-Review footer line
New API-Review label will be added to help identifying changes
needing API review. In first phase the label is non-blocking
and for information only, so it doesn't make sense to display
it in commit footer. If later the label is changed to blocking,
it might make sense to revert this change.

Change-Id: Ib7120088c995920d7e2b2ac54f70a8a21296bee1
Reviewed-by: Daniel Smith <Daniel.Smith@qt.io>