Validate commit message before staging
If a commit message contains extra line breaks or spaces after the
Change-Id footer, it can cause the change to get stuck in the
"integrating" status. This is due to limitations in
RevCommit.getFooterLines(), which fails to parse the Change-Id
correctly when the plugin attempts to cherry-pick the change to
a branch.

To prevent this, add a basic validation check before allowing the
change to be staged.

Task-number: QTQAINFRA-7223
Change-Id: I4a5265d7b20a854e3b547b54f10ebe1008769cec
Reviewed-by: Daniel Smith <daniel.smith@qt.io>