Serialize staging branch operations with per-branch locks
In the previous patch, making staging branch locks global
will introduce delays across the board. This patch introduces
per-branch locks to allow concurrent operations on different
staging branches while still preventing concurrent operations
on the same branch. While this now means that a user action
of clicking the "Stage" button on a change may incur a delay
or even a timeout if a long operation is underway on the branch,
it should now be safe. Further optimization using a
"queued for staging" state is possible to speed up
UI responsiveness, but not implemented in this patch.

Task-number: QTQAINFRA-7317
Change-Id: I47b8f3f600adddeb022496383aaacf9704fd805e
Reviewed-by: Jukka Jokiniva <jukka.jokiniva@qt.io>