ui: Add cancel confirmation to precheck dialog
This patch intercepts the Escape Key event in the precheck
dialog, as well as the cancel button click, to show a confirmation
dialog before closing and discarding the precheck dialog.

During the implementation of this change, it was discovered that the
precheck dialog left zombie elements in the DOM after being closed.
This patch also addresses that issue by ensuring proper cleanup of
the dialog elements on cancel. Implementation also required removal
of the default gerrit UI cancel handler to avoid bypassing
the confirmation dialog.

Task-number: QTQAINFRA-5878
Change-Id: If88bd512781ce7818b9fb90e99d957ae38f8f7dd
Reviewed-by: Jukka Jokiniva <jukka.jokiniva@qt.io>