ui: Use gerrit style variables where possible and change button order
The gerrit workflow plugin ui uses a bunch of hard-coded values
for styling which break dark-mode usage. In addition, the cancel
and confirm buttons are in the reverse order compared with default
gerrit dialogs.

This patch moves styling to use gerrit style variables where possible
and switches the button order to match gerrit defaults.

Fixes: QTQAINFRA-7494
Change-Id: Ib4b83e772da9cfb90464a49c7323f1a5dcc20b91
Reviewed-by: Jukka Jokiniva <jukka.jokiniva@qt.io>